Revising the layout method to include the edge weight. Currently revising the class compatibility.

+++ gila-core/src/unipg/gila/layout/single/SingleScaleLayout.java
@@ -24,6 +24,7 @@ import org.apache.giraph.graph.Vertex;
 import org.apache.giraph.worker.WorkerContext;
 import org.apache.giraph.worker.WorkerGlobalCommUsage;
 import org.apache.hadoop.io.FloatWritable;
+import org.apache.hadoop.io.IntWritable;
 import org.apache.hadoop.io.NullWritable;
 
 import unipg.gila.common.coordinatewritables.CoordinateWritable;
@@ -37,7 +38,7 @@ import unipg.gila.layout.LayoutRoutine.LayoutCCs;
 
 public class SingleScaleLayout {
 	
-	public static class Seeder extends AbstractSeeder<PartitionedLongWritable, CoordinateWritable, FloatWritable, LayoutMessage, LayoutMessage>{
+	public static class Seeder extends AbstractSeeder<PartitionedLongWritable, CoordinateWritable, IntWritable, LayoutMessage, LayoutMessage>{
 		
 		/* (non-Javadoc)
 		 * @see unipg.gila.layout.AbstractSeeder#initialize(org.apache.giraph.graph.GraphState, org.apache.giraph.comm.WorkerClientRequestProcessor, org.apache.giraph.graph.GraphTaskManager, org.apache.giraph.worker.WorkerGlobalCommUsage, org.apache.giraph.worker.WorkerContext)
@@ -45,8 +46,8 @@ public class SingleScaleLayout {
 		@Override
 		public void initialize(
 				GraphState graphState,
-				WorkerClientRequestProcessor<PartitionedLongWritable, CoordinateWritable, FloatWritable> workerClientRequestProcessor,
-				GraphTaskManager<PartitionedLongWritable, CoordinateWritable, FloatWritable> graphTaskManager,
+				WorkerClientRequestProcessor<PartitionedLongWritable, CoordinateWritable, IntWritable> workerClientRequestProcessor,
+				GraphTaskManager<PartitionedLongWritable, CoordinateWritable, IntWritable> graphTaskManager,
 				WorkerGlobalCommUsage workerGlobalCommUsage,
 				WorkerContext workerContext) {
 			super.initialize(graphState, workerClientRequestProcessor, graphTaskManager,
@@ -58,7 +59,7 @@ public class SingleScaleLayout {
 		 */
 		@Override
 		public void compute(
-				Vertex<PartitionedLongWritable, CoordinateWritable, FloatWritable> vertex,
+				Vertex<PartitionedLongWritable, CoordinateWritable, IntWritable> vertex,
 				Iterable<LayoutMessage> messages) throws IOException {
 			super.compute(vertex, messages);
 		}
@@ -79,7 +80,7 @@ public class SingleScaleLayout {
 	 * @author Alessio Arleo
 	 *
 	 */
-	public static class Propagator extends AbstractPropagator<PartitionedLongWritable, CoordinateWritable, FloatWritable, LayoutMessage, LayoutMessage>{
+	public static class Propagator extends AbstractPropagator<PartitionedLongWritable, CoordinateWritable, IntWritable, LayoutMessage, LayoutMessage>{
 
 
 		/* (non-Javadoc)
@@ -88,8 +89,8 @@ public class SingleScaleLayout {
 		@Override
 		public void initialize(
 				GraphState graphState,
-				WorkerClientRequestProcessor<PartitionedLongWritable, CoordinateWritable, FloatWritable> workerClientRequestProcessor,
-				GraphTaskManager<PartitionedLongWritable, CoordinateWritable, FloatWritable> graphTaskManager,
+				WorkerClientRequestProcessor<PartitionedLongWritable, CoordinateWritable, IntWritable> workerClientRequestProcessor,
+				GraphTaskManager<PartitionedLongWritable, CoordinateWritable, IntWritable> graphTaskManager,
 				WorkerGlobalCommUsage workerGlobalCommUsage,
 				WorkerContext workerContext) {
 			super.initialize(graphState, workerClientRequestProcessor, graphTaskManager,
@@ -101,7 +102,7 @@ public class SingleScaleLayout {
 		 */
 		@Override
 		public void compute(
-				Vertex<PartitionedLongWritable, CoordinateWritable, FloatWritable> vertex,
+				Vertex<PartitionedLongWritable, CoordinateWritable, IntWritable> vertex,
 				Iterable<LayoutMessage> messages) throws IOException {
 			super.compute(vertex, messages);
 		}
